chore(build): add TS support
