+++ lerna.json
@@ -13,7 +13,8 @@
   "packages": [
     "packages/eslint-plugin-pf-codemods",
     "packages/pf-codemods",
-    "packages/class-name-updater"
+    "packages/class-name-updater",
+    "packages/codemods"
   ],
   "version": "independent"
 }
