+++ packages/class-name-updater/src/utils.ts
@@ -1,0 +1,8 @@
+import { readdir }  from "fs/promises";
+
+export async function isDir(potentialDirPath: string) {
+  return readdir(potentialDirPath).then(
+    () => true,
+    () => false
+  );
+}
