--- packages/class-name-updater/utils.js
@@ -1,12 +1,0 @@
-const { readdir } = require("fs").promises;
-
-async function isDir(potentialDirPath) {
-  return readdir(potentialDirPath).then(
-    () => true,
-    () => false
-  );
-}
-
-module.exports = {
-  isDir,
-};
