+++ packages/eslint-plugin-pf-codemods/package.json
@@ -2,15 +2,27 @@
   "name": "@patternfly/eslint-plugin-pf-codemods",
   "version": "1.84.2",
   "description": "ESLint plugin for PF codemods",
-  "main": "./index.js",
+  "main": "./dist/js/index.js",
+  "module": "./dist/esm/index.js",
+  "types": "./dist/esm/index.d.ts",
   "author": "Red Hat",
   "license": "MIT",
   "publishConfig": {
     "access": "public"
   },
+  "scripts": {
+    "build": "yarn build:esm && yarn build:cjs",
+    "build:esm": "tsc --build --verbose ./tsconfig.json",
+    "build:cjs": "tsc --build --verbose ./tsconfig.cjs.json",
+    "test": "eslint . --ext .ts",
+    "clean": "rimraf ./dist"
+  },
   "peerDependencies": {
-    "@typescript-eslint/parser": "^5.58.0",
+    "@typescript-eslint/parser": ">=5.58.0",
     "eslint": ">=0.8.0",
-    "typescript": "^3.8.3"
-  }
+    "typescript": ">=3.8.3"
+  },
+  "devDependencies": {
+    "@types/eslint": "^8.56.0"
+  } 
 }
