+++ packages/eslint-plugin-pf-codemods/tsconfig.cjs.json
@@ -1,0 +1,9 @@
+{
+  "extends": "../../tsconfig.base.json",
+  "compilerOptions": {
+    "module": "commonjs",
+    "outDir": "./dist/js"
+  },
+  "include": ["src/**/*"],
+  "exclude": ["dist"]
+}
